+++ include/stlab/concurrency/default_executor.hpp
@@ -79,21 +79,7 @@ struct group_t {
     }
 };
 
-#if STLAB_BUILD_LIBRARY()
 group_t& group();
-#else
-inline auto group() -> group_t& {
-    // Use an immediately executed lambda to atomically register pre-exit handler
-    // and create the dispatch group.
-    static group_t g{[] {
-        at_pre_exit([]() noexcept { // <br>
-            dispatch_group_wait(g._group, DISPATCH_TIME_FOREVER);
-        });
-        return group_t{};
-    }()};
-    return g;
-}
-#endif
 
 template <executor_priority P = executor_priority::medium>
 struct executor_type {
@@ -456,19 +442,7 @@ public:
 /// Returns an instance of the task system singleton. An immediately executed lambda is used
 /// to register the the task system for tear down pre-exit in a thread safe manner.
 
-#if STLAB_BUILD_LIBRARY()
 priority_task_system& pts();
-#else
-inline priority_task_system& pts() {
-    // Uses the `nullptr` constructor with an immediate executed lambda to register the task
-    // system in a thread safe manner.
-    static priority_task_system only_task_system{[] {
-        at_pre_exit([]() noexcept { only_task_system.join(); });
-        return nullptr;
-    }()};
-    return only_task_system;
-}
-#endif
 
 #endif
